modifications to ESP32C3
