+++ uCNC/src/hal/boards/esp32/boardmap_core_c3.h
@@ -25,7 +25,7 @@ extern "C"
 #endif
 
 #ifndef MCU
-#define MCU MCU_ESP32C
+#define MCU MCU_ESP32C3
 #endif
 
 #ifndef BOARD_NAME
@@ -34,59 +34,62 @@ extern "C"
 
 // SAME AS GRBL for test purposes
 // Setup step pins
-#define STEP2_BIT 17 // assigns STEP2 pin
-#define STEP1_BIT 25 // assigns STEP1 pin
-#define STEP0_BIT 26 // assigns STEP0 pin
+#define STEP2_BIT 2 // assigns STEP2 pin
+#define STEP1_BIT 1 // assigns STEP1 pin
+#define STEP0_BIT 0 // assigns STEP0 pin
 
 // Setup dir pins
-#define DIR2_BIT 14 // assigns DIR2 pin
-#define DIR1_BIT 27 // assigns DIR1 pin
-#define DIR0_BIT 16 // assigns DIR0 pin
+#define DIR2_BIT 5 // assigns DIR2 pin
+#define DIR1_BIT 4 // assigns DIR1 pin
+#define DIR0_BIT 3 // assigns DIR0 pin
 
 // Setup control input pins
-#define ESTOP_BIT 2
-#define ESTOP_ISR
-#define ESTOP_PULLUP
-#define FHOLD_BIT 4
-#define FHOLD_ISR
-#define FHOLD_PULLUP
-#define CS_RES_BIT 36
-#define CS_RES_ISR
-#define CS_RES_PULLUP
+// #define ESTOP_BIT 6
+// #define ESTOP_ISR
+// #define ESTOP_PULLUP
+// #define FHOLD_BIT 7
+// #define FHOLD_ISR
+// #define FHOLD_PULLUP
+// #define CS_RES_BIT 8
+// #define CS_RES_ISR
+// #define CS_RES_PULLUP
 
 // Setup limit pins
-#define LIMIT_Z_BIT 19 // assigns LIMIT_Z pin
-#define LIMIT_Z_ISR		 // assigns LIMIT_Z ISR
-#define LIMIT_Y_BIT 5	 // assigns LIMIT_Y pin
-#define LIMIT_Y_ISR		 // assigns LIMIT_Y ISR
-#define LIMIT_X_BIT 13 // assigns LIMIT_X pin
-#define LIMIT_X_ISR		 // assigns LIMIT_X ISR
+// #define LIMIT_Z_BIT 9 // assigns LIMIT_Z pin
+// #define LIMIT_Z_ISR		 // assigns LIMIT_Z ISR
+// #define LIMIT_Y_BIT 10	 // assigns LIMIT_Y pin
+// #define LIMIT_Y_ISR		 // assigns LIMIT_Y ISR
+// #define LIMIT_X_BIT 11 // assigns LIMIT_X pin
+// #define LIMIT_X_ISR		 // assigns LIMIT_X ISR
 
 // Setup probe pin
-#define PROBE_BIT 39
+#define PROBE_BIT 10
 #define PROBE_ISR
 
 // Setup com pins
-#define RX_BIT 3
-#define TX_BIT 1
+#define RX_BIT 20
+#define TX_BIT 21
 #define RX_PULLUP
 	// only uncomment this if other port other then 0 is used
 	// #define UART_PORT 0
 
+#define USB_DP_BIT 18
+#define USB_DM_BIT 19
+
 	// Setup PWM
-#define PWM0_BIT 23 // assigns PWM0 pin
+#define PWM0_BIT 9 // assigns PWM0 pin
 #define PWM0_TIMER 0
 #define PWM0_CHANNEL 0
 
 // Setup generic IO Pins
 // spindle dir
-#define DOUT0_BIT 18
+#define DOUT0_BIT 8
 
 // coolant
-#define DOUT2_BIT 34
+// #define DOUT2_BIT 15
 
 // Stepper enable pin. For Grbl on Uno board a single pin is used
-#define STEP0_EN_BIT 12
+#define STEP0_EN_BIT 3
 
 	// Setup the Step Timer used has the heartbeat for µCNC
 	// Timer 1 is used by default
