+++ uCNC/src/hal/boards/esp32/esp32.ini
@@ -48,11 +48,6 @@ build_flags = ${common_esp32.build_flags} -D BOARDMAP=\"src/hal/boards/esp32/boa
 extends = common_esp32
 build_flags = ${common_esp32.build_flags} -D BOARDMAP=BOARDMAP_FILE("src/hal/boards/esp32/boardmap_esp32_shield_v3.h")
 
-[env:ESP32-C3]
-extends = common_esp32
-board = airm2m_core_esp32c3
-build_flags = ${common_esp32.build_flags} -D BOARDMAP=BOARDMAP_FILE("src/hal/boards/esp32/boardmap_core_c3.h")
-
 [common_esp32s]
 platform = espressif32
 framework = arduino
@@ -89,6 +84,38 @@ build_flags = ${common_esp32s.build_flags} -D BOARDMAP=BOARDMAP_FILE("src/hal/bo
 extends = common_esp32
 build_flags = ${common_esp32.build_flags} -DMCU=MCU_CUSTOM
 board = ${env.board}
+
+[common_esp32c]
+platform = espressif32
+framework = arduino
+board = airm2m_core_esp32c3
+bulid_unflags = -DARDUINO_USB_MODE
+build_flags = ${env.build_flags} -Wno-frame-address -ffunction-sections -fdata-sections -ggdb -Os -freorder-blocks -Wwrite-strings -fstack-protector -fstrict-volatile-bitfields -Wall -fno-jump-tables -fno-tree-switch-conversion -std=gnu++11 -fexceptions -MMD -c -DESP32C3 -DUSE_ARDUINO_EEPROM_LIBRARY -DUSE_ARDUINO_CDC
+board_build.f_flash = 80000000L
+board_build.f_cpu = 240000000L
+board_build.partitions = min_spiffs.csv
+monitor_filters=esp32_exception_decoder
+upload_speed = 921600
+upload_protocol = esptool
+board_build.filesystem = littlefs
+lib_deps = 
+    ${env.lib_deps}
+		https://github.com/Links2004/arduinoWebSockets/archive/refs/heads/master.zip
+		SPI
+		Wire
+		Wifi
+		WebServer
+		HTTPUpdatedServer
+		BluetoothSerial
+		Hash
+		LittleFS
+		EEPROM
+		; https://github.com/Paciente8159/uCNC-tinyusb.git#master
+
+[env:ESP32-C3]
+extends = common_esp32c
+board = airm2m_core_esp32c3
+build_flags = ${common_esp32c.build_flags} -D BOARDMAP=BOARDMAP_FILE("src/hal/boards/esp32/boardmap_core_c3.h")
 
 #################
 # ESPIDF builds #
